Don't run static analysis on app code
The static analysis is used to determine what "var"s need to be
automatically inserted by the linker at the package level.  But for app
code, we never actually insert these vars, since apps run in "global
namespace" mode.  So it's a waste of time to run the static analysis at
all.